Add watching and stopping on demand for the client
Client is now able to watch for changes in folder and show them on
the log.

It keeps on logging until it finds and specified (for now fixed)
stop mark (end.txt)

This stopping method is under construction as I'm unaware of the
way the files are constructed by the capturing algorithm, an once
I get the proper explanation of this, the stopping method might change.